requirements.txt
numpy==1.26.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
pillow==10.2.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
